- Fixed bug which prevented experiment from running when task or hardware test were in a subdirectory.
- Fixed bug which prevented task file from being copied to data folder if it was in a subdirectory.

- Added more detailed docstring to TaskSelectMenu class (renamed from menuSelect).

- Moved example tasks into 'example' subfolder.
